[svn] Fixed items status inside unversioned directory.
Summary:
Added special checks for items inside unversioned dirs.
Previously those items was indicated as normal revision with a capability to launch SVN commands on it (SVN Delete, for example). Those commands ends up with an error.

Unversioned item is a special case because in any other cases 'svn status' will mark every file or dir as added, removed, conflicted and so on.

Test Plan:
Inside a repo create dir with a following structure:
{F8225757}

1) Check any SVN actions available only for "dir", selecting other dirs and files causes no actions available.
2) Check items status showing correct.
3) Do 'SVN Add' for dir, 'SVN Revert' for dir2.
4) Check dir (with dir1.txt, file1.txt) and dir/dir1 (with file2.txt) and dir/dir1/dir2 has SVN Actions while dir/dir1/dir2/dir3 don't.
5) Check items status showing correct.

Reviewers: meven, elvisangelaccio, #dolphin

Reviewed By: meven, elvisangelaccio, #dolphin

Differential Revision: https://phabricator.kde.org/D28681