Implement mount ISO service plugin for Dolphin
The mountiso plugin adds a contextual menu entry for .iso files to
either mounting or unmounting the file as a loopback device depending on
the mount state.
This plugin uses udisksctl to either mount or unmount the iso file.
It also uses losetup to check if an iso file is already mounted

FEATURE: 175051
FIXED-IN: 20.08.0