[Mount ISO Action] Optimize getDeviceFromBackingFile
Only query for `StorageAccess` as only this device type can be mounted,
saves around 2/3 of needlessly created `Devices`, significantly speeding
up the lookup.

Also use `property()` instead of `allProperties()`, the latter of which
does an explicit `GetAll` DBus query. Effectively, will probably not do
much because it will be the first time the `Device` properties are
queried but it's still slightly convoluted.