Add Makefile Actions Plugin for Dolphin
Here is a Dolphin plugin that allows to make Makefile targets.

It automatically reads the Makefile database and populate a submenu in the context menu to allow building the targets. 

It can either open a Konsole window or run the command in the background and in that case it will only notify the user if the build failed. During the build, Dolphin is still usable but its cursor is changed to the busy one.

You can opt-in launching make commands in a terminal.