git: Allow pushing new branches
This commit makes sure the remote branch combobox always also contains
the name of the local branch. This way one can create a new branch and
push it, which then creates the new branch on the remote. The "Push"
confirmation button is then changed to "Push New Branch".

This commit also adds a warning box that appears when no remote
repository has been added yet and pushing is therefore not possible.